// src/utils/interfaces/SignUpValues.ts
/*
 *
 *  * Copyright (c) 2023 TechAxis.
 *  * All rights reserved.
 *  * Redistribution and use in source and binary forms, with or without modification, are not permitted.
 *
 */

export interface SignUpValues {
  firstName: string;
  lastName: string;
  email: string;
  // phoneNumber: string;
  password: string;
  confirmPassword?: string;
}